1.6.2 Health System updates

+++ src/main/java/com/mmiillkkaa/supernaturals/manager/DemonManager.java
@@ -369,7 +369,7 @@ public class DemonManager extends ClassManager {
 			return;
 		}
 
-		int health = player.getHealth();
+		double health = player.getHealth();
 		health += SNConfigHandler.demonHealing;
 		if (health > player.getMaxHealth()) {
 			health = player.getMaxHealth();
