+++ src/main/java/com/mmiillkkaa/supernaturals/manager/PriestManager.java
@@ -453,7 +453,7 @@ public class PriestManager extends HumanManager {
 				SuperNManager.sendMessage(snvictim, "You were healed by "
 						+ ChatColor.WHITE + snplayer.getName() + ChatColor.RED
 						+ "!");
-				int health = victim.getHealth()
+				double health = victim.getHealth()
 						+ SNConfigHandler.priestHealAmount;
 				if (health > victim.getMaxHealth()) {
 					health = victim.getMaxHealth();
